requirements.txt
streamlit>=1.33.0
selenium>=4.20.0
undetected-chromedriver>=3.5.5
pillow>=10.3.0
